requirements.txt
django
djangorestframework
requests
factory_boy
